remove annotationUtil.ts
